+++ docs/CHANGELOG.md
@@ -10,6 +10,7 @@ Version 1.8.1 (not released yet)
 - Text improvements to setup
 - Don't copy over the changelog during setup
 - Add `gamma` command
+- Fix holes in driller cockpit on a widescreen display
 
 Version 1.8.0 (released 2022-09-17)
 -----------------------------------